+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Removed
+
+- Dependency `tarampampam/wrappers-php` because this package was deprecated and removed
+
 ## v4.0.0
 
 ### Added